CSP-2591 Moved GetCourseLookupRequest from shared to FAT
