+++ src/FindApprenticeshipTraining/SFA.DAS.FindApprenticeshipTraining.UnitTests/Application/Services/WhenGettingStandardDetails.cs
@@ -14,6 +14,7 @@ using SFA.DAS.Apim.Shared.Interfaces;
 using SFA.DAS.Apim.Shared.Models;
 using SFA.DAS.SharedOuterApi.Types.Models;
 using SFA.DAS.Testing.AutoFixture;
+using SFA.DAS.FindApprenticeshipTraining.InnerApi.Requests;
 
 namespace SFA.DAS.FindApprenticeshipTraining.UnitTests.Application.Services;
