+++ src/FindApprenticeshipTraining/SFA.DAS.FindApprenticeshipTraining/Services/CachedStandardDetailsService.cs
@@ -1,10 +1,11 @@
 ﻿using System.Threading.Tasks;
 using SFA.DAS.Apim.Shared.Extensions;
+using SFA.DAS.Apim.Shared.Interfaces;
+using SFA.DAS.FindApprenticeshipTraining.InnerApi.Requests;
 using SFA.DAS.SharedOuterApi.Types.Configuration;
 using SFA.DAS.SharedOuterApi.Types.InnerApi.Requests.Courses;
 using SFA.DAS.SharedOuterApi.Types.InnerApi.Responses.Courses;
 using SFA.DAS.SharedOuterApi.Types.Interfaces;
-using SFA.DAS.Apim.Shared.Interfaces;
 
 namespace SFA.DAS.FindApprenticeshipTraining.Services;
